#15: Add XPCShell test skeleton
A duplicate of rpl's first patch in this [gist](https://gist.github.com/rpl/de10cdef052375961b392ced7909f3ff) that sets up most of the environment that we need to move our mochitest to an xpcshell test.

This test currently hangs when 'storageFront.listStores' is called. The offending code is inside of the indexedDB storage actor's 'preListStores' method. The next commit will address this issue.