#15: Create missing 'indexedDB' storage actor dir(s)
Now our xpcshell test passes! ('./mach test devtools/server/tests/unit/test_extension_storage_actor.js')

See [this comment](https://github.com/biancadanforth/bug_1542035/issues/15#issuecomment-489705375) for an explanation of this.

We will want to separate this commit from the other commits when pushing this to Phabricator for review, as it is potentially a temporary fix for an issue with the 'indexedDB' storage actor that is unrelated to our storage actor patch.