Cleanup preListStores method in storage actor
We no longer require 'window' objects for the extensionStorage storage actor (see #9), so we no longer need to iterate through 'this.windows' nor pass a 'window' argument into 'populateStoresForHost'.

Also make sure we actually have an extension to get storage data for, and that it is enabled before calling 'populateStoresForHost'.